+++ psutil/tests/test_posix.py
@@ -115,7 +115,10 @@ def ps_args(pid):
     field = "command"
     if AIX or SUNOS:
         field = "args"
-    return ps(field, pid)
+    out = ps(field, pid)
+    # observed on BSD + Github CI: '/usr/local/bin/python3 -E -O (python3.9)'
+    out = re.sub(r"\(python.*?\)$", "", out)
+    return out.strip()
 
 
 def ps_rss(pid):
@@ -347,11 +350,8 @@ class TestSystemAPIs(PsutilTestCase):
             for idx, u in enumerate(psutil.users()):
                 self.assertEqual(u.name, users[idx])
                 self.assertEqual(u.terminal, terminals[idx])
-                p = psutil.Process(u.pid)
-                # on macOS time is off by ~47 secs for some reason, but
-                # the next test against 'who' CLI succeeds
-                delta = 60 if MACOS else 1
-                self.assertAlmostEqual(u.started, p.create_time(), delta=delta)
+                if u.pid is not None:  # None on OpenBSD
+                    psutil.Process(u.pid)
 
     @retry_on_failure()
     def test_users_started(self):
@@ -415,7 +415,12 @@ class TestSystemAPIs(PsutilTestCase):
     @retry_on_failure()
     def test_disk_usage(self):
         def df(device):
-            out = sh("df -k %s" % device).strip()
+            try:
+                out = sh("df -k %s" % device).strip()
+            except RuntimeError as err:
+                if "device busy" in str(err).lower():
+                    raise self.skipTest("df returned EBUSY")
+                raise
             line = out.split('\n')[1]
             fields = line.split()
             total = int(fields[1]) * 1024